backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# File handling
pillow==11.3.0
python-magic==0.4.27

# Existing ML dependencies (from main requirements.txt)
opencv-python==4.12.0.88
pytesseract==0.3.13
numpy==1.26.3
torch==2.8.0
torchvision==0.23.0
scikit-learn==1.4.0

# Web scraping (existing)
selenium==4.35.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0

# Utilities
loguru==0.7.3
python-dotenv==1.0.1

# CORS
python-cors==1.0.0
